% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{countBarcodes}
\alias{countBarcodes}
\title{Count barcodes}
\usage{
countBarcodes(inp_fn, out_fn, wl_fn)
}
\arguments{
\item{inp_fn}{input FQ file}

\item{out_fn}{output counts file}

\item{wl_fn}{output counts file}
}
\description{
Count barcodes
}
