% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{extractMolecules}
\alias{extractMolecules}
\title{Extract putative DNA molecules from barcoded BAM file.}
\usage{
extractMolecules(bam_fn, tagname = "BX", max_dist = 500000L)
}
\arguments{
\item{bam_fn}{filename}

\item{tagname}{barcode tag name}

\item{max_dist}{maximum distance between reads in a molecule}
}
\description{
Extract putative DNA molecules from barcoded BAM file.
}
