% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{preprocessFastq}
\alias{preprocessFastq}
\title{Preprocess barcoded FASTQ files.}
\usage{
preprocessFastq(cts_inp_fn, fq1_inp_fn, fq2_inp_fn, fq1_out_fn, fq2_out_fn)
}
\arguments{
\item{cts_inp_fn}{input FQ file read 1}

\item{fq1_inp_fn}{input FQ file read 1}

\item{fq2_inp_fn}{input FQ file read 2}

\item{fq1_out_fn}{output FQ file read 1}

\item{fq2_out_fn}{output FQ file read 2}
}
\description{
Preprocess barcoded FASTQ files.
}
